# perf: Faster global settings lookup