--- a/migrations/frontend/1528395925_settings_global_index.down.sql
+++ b/migrations/frontend/1528395925_settings_global_index.down.sql
@@ -1,0 +1,5 @@
+BEGIN;
+
+DROP INDEX IF EXISTS settings_global_id;
+
+COMMIT;