--- a/migrations/frontend/1528395925_settings_global_index.up.sql
+++ b/migrations/frontend/1528395925_settings_global_index.up.sql
@@ -1,0 +1,5 @@
+BEGIN;
+
+CREATE INDEX IF NOT EXISTS settings_global_id ON settings (id DESC) WHERE user_id IS NULL AND org_id IS NULL;
+
+COMMIT;